requirements.txt
customtkinter
requests
Pillow
opencv-python
numpy
Flask